feat: Update link titles and URLs in features component
This commit updates the link titles and URLs in the features component to null. Previously, the links were pointing to specific URLs, but now they are set to null. This change was made to remove the hardcoded URLs and allow for dynamic link generation based on the specific use case. This improvement enhances the flexibility and reusability of the features component.