feat: Update Google Analytics email and campaign ID query parameters
This commit updates the query parameters used to retrieve the email and campaign ID in the GoogleAnalytics.tsx file. The email query parameter has been changed from "email" to "e", and the campaign ID query parameter has been changed from "cId" to "cid". This change ensures that the correct query parameters are used to retrieve the email and campaign ID values, improving the accuracy of the data tracked by Google Analytics.